Create package for image utility functions
