+++ examples/gradcam.py
@@ -6,6 +6,7 @@ import torch
 from torchvision import models
 
 from pytorch_grad_cam.gradcam import GradCam, preprocess_image, GuidedBackpropReLUModel
+from pytorch_grad_cam.utils.image import show_cam_on_image, deprocess_image
 
 
 def get_args():
@@ -22,24 +23,6 @@ def get_args():
         print('Using CPU for computation')
 
     return args
-
-
-def deprocess_image(img):
-    """ see https://github.com/jacobgil/keras-grad-cam/blob/master/grad-cam.py#L65 """
-    img = img - np.mean(img)
-    img = img / (np.std(img) + 1e-5)
-    img = img * 0.1
-    img = img + 0.5
-    img = np.clip(img, 0, 1)
-    return np.uint8(img * 255)
-
-
-def show_cam_on_image(img, mask):
-    heatmap = cv2.applyColorMap(np.uint8(255 * mask), cv2.COLORMAP_JET)
-    heatmap = np.float32(heatmap) / 255
-    cam = heatmap + np.float32(img)
-    cam = cam / np.max(cam)
-    return np.uint8(255 * cam)
 
 
 if __name__ == '__main__':
