+++ pytorch_grad_cam/utils/__init__.py
@@ -1,0 +1,1 @@
+from pytorch_grad_cam.utils.image import deprocess_image
